+++ src/client/snd_wavelet.c
@@ -37,18 +37,18 @@ If you have questions concerning this license or the applicable additional terms
 
 #include "snd_local.h"
 
-long myftol( float f );
+int myftol( float f );
 
 #define C0 0.4829629131445341
 #define C1 0.8365163037378079
 #define C2 0.2241438680420134
 #define C3 -0.1294095225512604
 
-void daub4( float b[], unsigned long n, int isign ) {
+void daub4( float b[], unsigned int n, int isign ) {
 	float wksp[4097];
 	float   *a = b - 1;                     // numerical recipies so a[1] = b[0]
 
-	unsigned long nh,nh1,i,j;
+	unsigned int nh,nh1,i,j;
 
 	if ( n < 4 ) {
 		return;
@@ -75,8 +75,8 @@ void daub4( float b[], unsigned long n, int isign ) {
 	}
 }
 
-void wt1( float a[], unsigned long n, int isign ) {
-	unsigned long nn;
+void wt1( float a[], unsigned int n, int isign ) {
+	unsigned int nn;
 	int inverseStartLength = n / 4;
 	if ( n < inverseStartLength ) {
 		return;
@@ -101,7 +101,7 @@ static unsigned char numBits[] = {
 };
 
 byte MuLawEncode( short s ) {
-	unsigned long adjusted;
+	unsigned int adjusted;
 	byte sign, exponent, mantissa;
 
 	sign = ( s < 0 ) ? 0 : 0x80;
@@ -109,7 +109,7 @@ byte MuLawEncode( short s ) {
 	if ( s < 0 ) {
 		s = -s;
 	}
-	adjusted = (long)s << ( 16 - sizeof( short ) * 8 );
+	adjusted = (int)s << ( 16 - sizeof( short ) * 8 );
 	adjusted += 128L + 4L;
 	if ( adjusted > 32767 ) {
 		adjusted = 32767;
@@ -120,7 +120,7 @@ byte MuLawEncode( short s ) {
 }
 
 short MuLawDecode( byte uLaw ) {
-	signed long adjusted;
+	int adjusted;
 	byte exponent, mantissa;
 
 	uLaw = ~uLaw;
